requirements.txt
openai>=1.0.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
pillow>=9.0.0
requests>=2.28.0
tqdm>=4.64.0